+++ python/core/auto_generated/symbology/qgssymbollayer.sip.in
@@ -660,7 +660,7 @@ This id has to be unique in the whole project
 .. versionadded:: 3.30
 %End
 
-    const QString &id() const;
+    QString id() const;
 %Docstring
 Returns symbol layer identifier
 This id is unique in the whole project
