+++ python/core/auto_generated/symbology/qgssymbollayerutils.sip.in
@@ -1036,17 +1036,6 @@ The method makes approximations and can modify ``angle`` in order to generate th
 .. versionadded:: 3.30
 %End
 
- static void fixOldSymbolLayerReferences( const QMap<QString, QgsMapLayer *> &mapLayers );
-%Docstring
-:py:class:`QgsSymbolLayerReference` uses :py:class:`QgsSymbolLayer` unique uuid identifier since QGIS 3.30, instead of the symbol
-key (rule for :py:class:`QgsRuleBasedRenderer` for instance) and index path, so this method migrates ``mapLayers`` old references
-to new ones.
-
-.. versionadded:: 3.30
-
-.. deprecated:: QGIS 3.30
-%End
-
     static void clearSymbolLayerIds( QgsSymbol *symbol );
 %Docstring
 Remove recursively unique id from all ``symbol`` symbol layers and set an empty string instead
