+++ src/core/project/qgsprojectfiletransform.h
@@ -25,6 +25,7 @@
 
 
 class QgsRasterLayer;
+class QgsMapLayer;
 
 /**
  * \ingroup core
@@ -76,6 +77,15 @@ class CORE_EXPORT QgsProjectFileTransform
      */
     QgsProjectVersion currentVersion() const;
 
+    /**
+     * QgsSymbolLayerReference uses QgsSymbolLayer unique uuid identifier since QGIS 3.30, instead of the symbol
+     * key (rule for QgsRuleBasedRenderer for instance) and index path, so this method migrates \a mapLayers old references
+     * to new ones.
+     * \since QGIS 3.30
+     * \deprecated since QGIS 3.30
+     */
+    Q_DECL_DEPRECATED static void fixOldSymbolLayerReferences( const QMap<QString, QgsMapLayer *> &mapLayers );
+
   private:
 
     QDomDocument mDom;
@@ -84,4 +94,3 @@ class CORE_EXPORT QgsProjectFileTransform
 
 
 #endif //QGSPROJECTFILETRANSFORM_H
-
