+++ src/core/qgsrendercontext.h
@@ -156,7 +156,7 @@ class CORE_EXPORT QgsRenderContext : public QgsTemporalRangeObject
      * \since QGIS 3.12
      * \deprecated since QGIS 3.30 and replaced with setDisabledSymbolLayersV2
      */
-    Q_DECL_DEPRECATED void setDisabledSymbolLayers( const QSet<const QgsSymbolLayer *> &symbolLayers );
+    Q_DECL_DEPRECATED void setDisabledSymbolLayers( const QSet<const QgsSymbolLayer *> &symbolLayers ) SIP_DEPRECATED;
 
     /**
      * When rendering a map layer in a second pass (for selective masking),
@@ -179,7 +179,7 @@ class CORE_EXPORT QgsRenderContext : public QgsTemporalRangeObject
      * \since QGIS 3.12
      * \deprecated since QGIS 3.30 and replaced with disabledSymbolLayersV2
      */
-    Q_DECL_DEPRECATED QSet<const QgsSymbolLayer *> disabledSymbolLayers() const;
+    Q_DECL_DEPRECATED QSet<const QgsSymbolLayer *> disabledSymbolLayers() const SIP_DEPRECATED;
 
     /**
      * When rendering a map layer in a second pass (for selective masking),
@@ -190,7 +190,7 @@ class CORE_EXPORT QgsRenderContext : public QgsTemporalRangeObject
      * \see isSymbolLayerEnabled()
      * \since QGIS 3.30
      */
-    const QSet<QString> &disabledSymbolLayersV2() const;
+    QSet<QString> disabledSymbolLayersV2() const;
 
     /**
      * When rendering a map layer in a second pass (for selective masking),
