+++ src/core/symbology/qgssymbollayerutils.h
@@ -930,15 +930,6 @@ class CORE_EXPORT QgsSymbolLayerUtils
     static QSize tileSize( int width, int height, double &angleRad SIP_INOUT );
 
     /**
-     * QgsSymbolLayerReference uses QgsSymbolLayer unique uuid identifier since QGIS 3.30, instead of the symbol
-     * key (rule for QgsRuleBasedRenderer for instance) and index path, so this method migrates \a mapLayers old references
-     * to new ones.
-     * \since QGIS 3.30
-     * \deprecated since QGIS 3.30
-     */
-    Q_DECL_DEPRECATED static void fixOldSymbolLayerReferences( const QMap<QString, QgsMapLayer *> &mapLayers );
-
-    /**
      * Remove recursively unique id from all \a symbol symbol layers and set an empty string instead
      * \since QGIS 3.30
      */
