+++ tests/src/python/test_selective_masking.py
@@ -16,8 +16,8 @@ import os
 import subprocess
 
 import qgis  # NOQA
-from qgis.PyQt.QtCore import QDir, QRectF, QSize, Qt
-from qgis.PyQt.QtGui import QColor, QImage, QPainter
+from qgis.PyQt.QtCore import QDir, QRectF, QSize, Qt, QUrl, QUuid
+from qgis.PyQt.QtGui import QColor, QImage, QPainter, QDesktopServices
 from qgis.core import (
     Qgis,
     QgsCoordinateReferenceSystem,
@@ -36,6 +36,7 @@ from qgis.core import (
     QgsOuterGlowEffect,
     QgsPalLayerSettings,
     QgsProject,
+    QgsProjectFileTransform,
     QgsProperty,
     QgsRectangle,
     QgsRenderChecker,
@@ -46,6 +47,7 @@ from qgis.core import (
     QgsSymbolLayerUtils,
     QgsUnitTypes,
     QgsWkbTypes,
+    QgsFontUtils
 )
 from qgis.testing import start_app, unittest
 
@@ -181,9 +183,6 @@ class TestSelectiveMasking(unittest.TestCase):
         - parallel rendering, with cache (rendered two times)
         - sequential rendering, with cache (rendered two times)
         """
-
-        return True
-
         for do_parallel in [False, True]:
             for use_cache in [False, True]:
                 print("=== parallel", do_parallel, "cache", use_cache)
@@ -318,7 +317,7 @@ class TestSelectiveMasking(unittest.TestCase):
         self.assertEqual([slRef.symbolLayerId() for slRef in self.polys_layer.labeling().settings().format().mask().maskedSymbolLayers()],
                          [slRef.symbolLayerId() for slRef in oldMaskRefs])
 
-        QgsSymbolLayerUtils.fixOldSymbolLayerReferences(QgsProject.instance().mapLayers())
+        QgsProjectFileTransform.fixOldSymbolLayerReferences(QgsProject.instance().mapLayers())
 
         self.assertEqual([QUuid(slRef.symbolLayerIdV2()).isNull() for slRef in self.polys_layer.labeling().settings().format().mask().maskedSymbolLayers()],
                          [False, False, False])
@@ -347,7 +346,7 @@ class TestSelectiveMasking(unittest.TestCase):
         self.assertEqual([slRef.symbolLayerId() for slRef in self.points_layer.renderer().symbol().symbolLayers()[1].masks()],
                          [slRef.symbolLayerId() for slRef in oldMaskRefs])
 
-        QgsSymbolLayerUtils.fixOldSymbolLayerReferences(QgsProject.instance().mapLayers())
+        QgsProjectFileTransform.fixOldSymbolLayerReferences(QgsProject.instance().mapLayers())
 
         self.assertEqual([QUuid(slRef.symbolLayerIdV2()).isNull() for slRef in self.points_layer.renderer().symbol().symbolLayers()[1].masks()],
                          [False])
@@ -1177,9 +1176,18 @@ class TestSelectiveMasking(unittest.TestCase):
         layer = QgsProject.instance().mapLayersByName('line_pattern_fill')[0]
         self.assertTrue(layer)
 
+        self.assertTrue(len(layer.labeling().subProviders()), 1)
+        settings = layer.labeling().settings()
+        fmt = settings.format()
+        fmt.setFont(QgsFontUtils.getStandardTestFont("Bold"))
+        fmt.setSize(30)
+        fmt.setSizeUnit(QgsUnitTypes.RenderPoints)
+        settings.setFormat(fmt)
+        layer.labeling().setSettings(settings)
+
         map_settings = QgsMapSettings()
         crs = QgsCoordinateReferenceSystem('epsg:4326')
-        extent = QgsRectangle(-0.972, -1.966, 1.58, 1.806)
+        extent = QgsRectangle(0, -1, 0.5, 0.8)
         map_settings.setBackgroundColor(QColor(152, 219, 249))
         map_settings.setOutputSize(QSize(420, 280))
         map_settings.setOutputDpi(72)
@@ -1200,6 +1208,15 @@ class TestSelectiveMasking(unittest.TestCase):
 
         layer = QgsProject.instance().mapLayersByName('line_pattern_fill')[0]
         self.assertTrue(layer)
+
+        self.assertTrue(len(layer.labeling().subProviders()), 1)
+        settings = layer.labeling().settings()
+        fmt = settings.format()
+        fmt.setFont(QgsFontUtils.getStandardTestFont("Bold"))
+        fmt.setSize(9)
+        fmt.setSizeUnit(QgsUnitTypes.RenderPoints)
+        settings.setFormat(fmt)
+        layer.labeling().setSettings(settings)
 
         map_settings = QgsMapSettings()
         crs = QgsCoordinateReferenceSystem('epsg:4326')
@@ -1223,6 +1240,15 @@ class TestSelectiveMasking(unittest.TestCase):
 
         layer = QgsProject.instance().mapLayersByName('point_pattern_fill')[0]
         self.assertTrue(layer)
+
+        self.assertTrue(len(layer.labeling().subProviders()), 1)
+        settings = layer.labeling().settings()
+        fmt = settings.format()
+        fmt.setFont(QgsFontUtils.getStandardTestFont("Bold"))
+        fmt.setSize(9)
+        fmt.setSizeUnit(QgsUnitTypes.RenderPoints)
+        settings.setFormat(fmt)
+        layer.labeling().setSettings(settings)
 
         map_settings = QgsMapSettings()
         crs = QgsCoordinateReferenceSystem('epsg:4326')
@@ -1248,6 +1274,15 @@ class TestSelectiveMasking(unittest.TestCase):
         layer = QgsProject.instance().mapLayersByName('centroid_fill')[0]
         self.assertTrue(layer)
 
+        self.assertTrue(len(layer.labeling().subProviders()), 1)
+        settings = layer.labeling().settings()
+        fmt = settings.format()
+        fmt.setFont(QgsFontUtils.getStandardTestFont("Bold"))
+        fmt.setSize(9)
+        fmt.setSizeUnit(QgsUnitTypes.RenderPoints)
+        settings.setFormat(fmt)
+        layer.labeling().setSettings(settings)
+
         map_settings = QgsMapSettings()
         crs = QgsCoordinateReferenceSystem('epsg:4326')
         extent = QgsRectangle(-1.0073971192118132, -0.7875782447946843, 0.87882587741257345, 0.51640826470600099)
@@ -1270,6 +1305,15 @@ class TestSelectiveMasking(unittest.TestCase):
 
         layer = QgsProject.instance().mapLayersByName('random_generator_fill')[0]
         self.assertTrue(layer)
+
+        self.assertTrue(len(layer.labeling().subProviders()), 1)
+        settings = layer.labeling().settings()
+        fmt = settings.format()
+        fmt.setFont(QgsFontUtils.getStandardTestFont("Bold"))
+        fmt.setSize(9)
+        fmt.setSizeUnit(QgsUnitTypes.RenderPoints)
+        settings.setFormat(fmt)
+        layer.labeling().setSettings(settings)
 
         map_settings = QgsMapSettings()
         crs = QgsCoordinateReferenceSystem('epsg:4326')
